Moved CSS files to correct folder
